Order

+++ Customer-Service/controllers/ticket_controller.js
@@ -7,7 +7,10 @@ module.exports = {
   createTicket(req, res, next) {
     const customerId = req.customerId;
     const ticket = req.body;
-
+    // if there is no customerId in the token, return error
+    if (!customerId) return res.status(401).send({ error: "Unauthorized" });
+    // if there is no title or description, return error
+    if (!ticket.title || !ticket.description) return res.status(422).send({ error: "Title and description are required." });
     rabbitMQManager.addMessage(
       `INSERT INTO Tickets (title, description, status, customerId, messages) VALUES ('${ticket.title}', '${ticket.description}', 'Open', '${customerId}', '[]')`
     );
@@ -24,6 +27,14 @@ module.exports = {
       return res.status(422).send({ error: "No tickets found." });
     } else {
       try {
+        // remove tickets that do not belong to the customer if the customer is not a service agent
+        if (req.role != "Service Agent") {
+          tickets.forEach((ticket) => {
+            if (ticket.customerId != req.customerId) {
+              tickets.splice(tickets.indexOf(ticket), 1);
+            }
+          });
+        }
         res.status(200).send(tickets);
       } catch (error) {
         console.log(error);
@@ -37,6 +48,7 @@ module.exports = {
     if (!ticket) {
       return res.status(422).send({ error: "Ticket does not exist." });
     } else {
+      if (req.role != "Service Agent" && req.customerId != ticket.customerId) return res.status(401).send({ error: "Unauthorized" });
       res.status(200).send(ticket);
     }
   },
@@ -48,6 +60,8 @@ module.exports = {
         customerId: customerId,
       },
     });
+    // check if customerId is the same as the one in the token or if the user is a service agent
+    if (req.role != "Service Agent" && req.customerId != customerId) return res.status(401).send({ error: "Unauthorized" });
     if (!tickets) {
       return res.status(422).send({ error: "No tickets found." });
     } else {
@@ -66,6 +80,14 @@ module.exports = {
         status: status,
       },
     });
+    // remove tickets that do not belong to the customer if the customer is not a service agent
+    if (req.role != "Service Agent") {
+      tickets.forEach((ticket) => {
+        if (ticket.customerId != req.customerId) {
+          tickets.splice(tickets.indexOf(ticket), 1);
+        }
+      });
+    }
     if (!tickets) {
       return res.status(422).send({ error: "No tickets found." });
     } else {
@@ -83,37 +105,26 @@ module.exports = {
     if (!ticket) {
       return res.status(422).send({ error: "Ticket does not exist." });
     } else {
-      await Ticket.destroy({
-        where: {
-          id: ticketId,
-        },
-      });
+      if (req.role != "Service Agent") return res.status(401).send({ error: "Unauthorized" });
       rabbitMQManager.addMessage(`DELETE FROM Tickets WHERE id = ${ticketId}`);
       res.status(200).send({ message: "Ticket deleted." });
     }
   },
 
-  async updateTicket(req, res, next) {
+  async updateStatus(req, res, next) {
     const ticketId = req.params.id;
     const ticket = await Ticket.findByPk(ticketId);
     if (!ticket) {
       return res.status(422).send({ error: "Ticket does not exist." });
     } else {
-      await Ticket.update(
-        {
-          title: req.body.title,
-          description: req.body.description,
-          status: req.body.status,
-        },
-        {
-          where: {
-            id: ticketId,
-          },
-        }
-      );
-      rabbitMQManager.addMessage(
-        `UPDATE Tickets SET title = '${req.body.title}', description = '${req.body.description}', status = '${req.body.status}' WHERE id = ${ticketId}`
-      );
+      if (req.role != "Service Agent") return res.status(401).send({ error: "Unauthorized" });
+      // if no status, return error
+      if (!req.body.status) return res.status(422).send({ error: "Status is required." });
+      // only accept open or closed as status (case insensitive)
+      if (req.body.status.toLowerCase() != "open" && req.body.status.toLowerCase() != "closed") return res.status(422).send({ error: "Invalid status." });
+      // make first letter uppercase
+      req.body.status = req.body.status.charAt(0).toUpperCase() + req.body.status.slice(1);
+      rabbitMQManager.addMessage(`UPDATE Tickets SET status = '${req.body.status}' WHERE id = ${ticketId}`);
       res.status(200).send({ message: "Ticket updated." });
     }
   },
@@ -121,24 +132,40 @@ module.exports = {
   async addComment(req, res, next) {
     const ticketId = req.params.id;
     const ticket = await Ticket.findByPk(ticketId);
+    const message = req.body.message;
     if (!ticket) {
       return res.status(422).send({ error: "Ticket does not exist." });
     } else {
-      await Ticket.update(
-        {
-          messages: req.body.messages,
-        },
-        {
-          where: {
-            id: ticketId,
-          },
-        }
-      );
+      // check if ticket belongs to customer or if role is service agent in one line
+      if (ticket.customerId != req.customerId && req.role != "Service Agent") return res.status(401).send({ error: "Unauthorized" });
+      // if no message, return error
+      if (!message) return res.status(422).send({ error: "No message." });
+      // add message to current messages
+      const messages = ticket.messages;
+      messages.push({
+        author: req.role,
+        message: message,
+        date: new Date().toISOString().slice(0, 19).replace("T", " "),
+      });
+      // parse messages as json
+      const messagesJson = JSON.stringify(messages);
       rabbitMQManager.addMessage(
-        `UPDATE Tickets SET messages = '${req.body.messages}' WHERE id = ${ticketId}`
+        `UPDATE Tickets SET messages = '${messagesJson}' WHERE id = ${ticketId}`
       );
       res.status(200).send({ message: "Comment added." });
     }
+  },
+
+  // simple method to generate service agent token
+  generateToken(req, res, next) {
+    const token = jwt.sign(
+      {
+        role: "Service Agent",
+      },
+      config.secret,
+      { expiresIn: "7d" }
+    );
+    res.status(200).json({ token: token });
   },
 
   validateToken(req, res, next) {
@@ -152,9 +179,10 @@ module.exports = {
           return res.status(401).json({ message: "Invalid token" });
         } else {
           // check if token has customerId
-          if (!decoded.sub)
+          if (!decoded.sub && !decoded.role)
             return res.status(401).json({ message: "Invalid token" });
           req.customerId = decoded.sub;
+          req.role = decoded.role || "Customer";
           next();
         }
       });
